# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * A pet tag profile. Never exposes the PIN.
 */
export interface Pet {
  /** The tag's unique id, embedded in the NFC tag's URL. */
  id: string;
  /** Whether an owner has filled in and locked this profile yet. */
  claimed: boolean;
  /** @nullable */
  name: string | null;
  /**
     * e.g. dog, cat
     * @nullable
     */
  species: string | null;
  /** @nullable */
  breed: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  ownerName: string | null;
  /** @nullable */
  ownerPhone: string | null;
  /**
     * Object storage path for the pet's photo, e.g. /objects/uploads/uuid.
     * @nullable
     */
  photoObjectPath: string | null;
  createdAt: string;
  /** @nullable */
  claimedAt: string | null;
}

export interface PetClaimInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  species: string;
  breed?: string;
  description?: string;
  ownerName?: string;
  /** @minLength 1 */
  ownerPhone: string;
  photoObjectPath?: string;
  /**
     * A 4-8 digit/character PIN the owner will use to unlock editing later.
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}

export interface PetUpdateInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  species?: string;
  breed?: string;
  description?: string;
  ownerName?: string;
  /** @minLength 1 */
  ownerPhone?: string;
  photoObjectPath?: string;
}

export interface PetPinInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}

export interface PinCheckResult {
  valid: boolean;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}