# lib/api-zod/src/generated/types/pet.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * A pet tag profile. Never exposes the PIN.
 */
export interface Pet {
  /** The tag's unique id, embedded in the NFC tag's URL. */
  id: string;
  /** Whether an owner has filled in and locked this profile yet. */
  claimed: boolean;
  /** @nullable */
  name: string | null;
  /**
     * e.g. dog, cat
     * @nullable
     */
  species: string | null;
  /** @nullable */
  breed: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  ownerName: string | null;
  /** @nullable */
  ownerPhone: string | null;
  /**
     * Object storage path for the pet's photo, e.g. /objects/uploads/uuid.
     * @nullable
     */
  photoObjectPath: string | null;
  createdAt: Date;
  /** @nullable */
  claimedAt: Date | null;
}