# lib/api-zod/src/generated/types/petClaimInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PetClaimInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  species: string;
  breed?: string;
  description?: string;
  ownerName?: string;
  /** @minLength 1 */
  ownerPhone: string;
  photoObjectPath?: string;
  /**
     * A 4-8 digit/character PIN the owner will use to unlock editing later.
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}