# lib/api-zod/src/generated/types/petPinInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PetPinInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}