# lib/api-zod/src/generated/types/petUpdateInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PetUpdateInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  species?: string;
  breed?: string;
  description?: string;
  ownerName?: string;
  /** @minLength 1 */
  ownerPhone?: string;
  photoObjectPath?: string;
}